renamed the loggers to DebugLogger and ScenarioLog, also updated include paths and BUILD files
