+++ invertedai_cpp/invertedai/debug_logger.h
@@ -1,0 +1,35 @@
+#include <iostream>
+#include <vector>
+#include <string>
+
+namespace invertedai {
+class DebugLogger {
+    private:
+        std::vector<std::string> loc_requests_;
+        std::vector<std::string> loc_responses_;
+        std::vector<std::string> loc_request_timestamps_;
+        std::vector<std::string> loc_response_timestamps_;
+
+        std::vector<std::string> init_requests_;
+        std::vector<std::string> init_responses_;
+        std::vector<std::string> init_request_timestamps_;
+        std::vector<std::string> init_response_timestamps_;
+
+        std::vector<std::string> drive_requests_;
+        std::vector<std::string> drive_responses_;
+        std::vector<std::string> drive_request_timestamps_;
+        std::vector<std::string> drive_response_timestamps_;
+
+        std::string get_current_time_UTC_();
+
+    public:
+
+        void append_request(const std::string &req, const std::string &mode);
+
+        void append_response(const std::string &res, const std::string &mode);
+
+        void write_scenario_log(const std::string &dir_path,const std::string &log_path);
+
+        void write_log_to_file(const std::string &file_path, const bool &is_scenario_log);
+};
+}
